Split test_macros up
