+++ tests/unit/test_macros/test_macros.py
@@ -29,10 +29,6 @@ from evdev.ecodes import (
     REL_Y,
     REL_HWHEEL,
     REL_HWHEEL_HI_RES,
-    KEY_A,
-    KEY_B,
-    KEY_C,
-    KEY_E,
 )
 
 from inputremapper.configs.keyboard_layout import keyboard_layout
@@ -47,40 +43,6 @@ from tests.unit.test_macros.macro_test_base import MacroTestBase, DummyMapping
 
 @test_setup
 class TestMacros(MacroTestBase):
-    async def test_run_plus_syntax(self):
-        macro = Parser.parse("a + b + c + d", self.context, DummyMapping)
-
-        macro.press_trigger()
-        asyncio.ensure_future(macro.run(self.handler))
-        await asyncio.sleep(0.2)
-        self.assertTrue(macro.tasks[0].is_holding())
-
-        # starting from the left, presses each one down
-        self.assertEqual(self.result[0], (EV_KEY, keyboard_layout.get("a"), 1))
-        self.assertEqual(self.result[1], (EV_KEY, keyboard_layout.get("b"), 1))
-        self.assertEqual(self.result[2], (EV_KEY, keyboard_layout.get("c"), 1))
-        self.assertEqual(self.result[3], (EV_KEY, keyboard_layout.get("d"), 1))
-
-        # and then releases starting with the previously pressed key
-        macro.release_trigger()
-        await asyncio.sleep(0.2)
-        self.assertFalse(macro.tasks[0].is_holding())
-        self.assertEqual(self.result[4], (EV_KEY, keyboard_layout.get("d"), 0))
-        self.assertEqual(self.result[5], (EV_KEY, keyboard_layout.get("c"), 0))
-        self.assertEqual(self.result[6], (EV_KEY, keyboard_layout.get("b"), 0))
-        self.assertEqual(self.result[7], (EV_KEY, keyboard_layout.get("a"), 0))
-
-    async def test_child_macro_count(self):
-        # It correctly keeps track of child-macros for both positional and keyword-args
-        macro = Parser.parse(
-            "hold(macro=hold(hold())).repeat(1, macro=repeat(1, hold()))",
-            self.context,
-            DummyMapping,
-            True,
-        )
-        self.assertEqual(self.count_child_macros(macro), 4)
-        self.assertEqual(self.count_tasks(macro), 6)
-
     async def test_0(self):
         macro = Parser.parse("key(1)", self.context, DummyMapping, True)
         one_code = keyboard_layout.get("1")
@@ -175,218 +137,6 @@ class TestMacros(MacroTestBase):
                 (EV_KEY, code_b, 0),
             ],
         )
-
-    async def test_hold_variable(self):
-        code_a = keyboard_layout.get("a")
-        macro = Parser.parse("set(foo, a).hold($foo)", self.context, DummyMapping)
-        await macro.run(self.handler)
-        self.assertListEqual(
-            self.result,
-            [
-                (EV_KEY, code_a, 1),
-                (EV_KEY, code_a, 0),
-            ],
-        )
-
-    async def test_hold_keys(self):
-        macro = Parser.parse(
-            "set(foo, b).hold_keys(a, $foo, c)", self.context, DummyMapping
-        )
-        # press first
-        macro.press_trigger()
-        # then run, just like how it is going to happen during runtime
-        asyncio.ensure_future(macro.run(self.handler))
-
-        code_a = keyboard_layout.get("a")
-        code_b = keyboard_layout.get("b")
-        code_c = keyboard_layout.get("c")
-
-        await asyncio.sleep(0.2)
-        self.assertListEqual(
-            self.result,
-            [
-                (EV_KEY, code_a, 1),
-                (EV_KEY, code_b, 1),
-                (EV_KEY, code_c, 1),
-            ],
-        )
-
-        macro.release_trigger()
-
-        await asyncio.sleep(0.2)
-        self.assertListEqual(
-            self.result,
-            [
-                (EV_KEY, code_a, 1),
-                (EV_KEY, code_b, 1),
-                (EV_KEY, code_c, 1),
-                (EV_KEY, code_c, 0),
-                (EV_KEY, code_b, 0),
-                (EV_KEY, code_a, 0),
-            ],
-        )
-
-    async def test_hold_keys_broken(self):
-        # Won't run any of the keys when one of them is invalid
-        macro = Parser.parse(
-            "set(foo, broken).hold_keys(a, $foo, c)", self.context, DummyMapping
-        )
-        # press first
-        macro.press_trigger()
-        # then run, just like how it is going to happen during runtime
-        asyncio.ensure_future(macro.run(self.handler))
-        await asyncio.sleep(0.2)
-        self.assertListEqual(self.result, [])
-        macro.release_trigger()
-        await asyncio.sleep(0.2)
-        self.assertListEqual(self.result, [])
-
-    async def test_hold(self):
-        # repeats key(a) as long as the key is held down
-        macro = Parser.parse("key(1).hold(key(a)).key(3)", self.context, DummyMapping)
-
-        """down"""
-
-        macro.press_trigger()
-        await asyncio.sleep(0.05)
-        self.assertTrue(macro.tasks[1].is_holding())
-
-        macro.press_trigger()  # redundantly calling doesn't break anything
-        asyncio.ensure_future(macro.run(self.handler))
-        await asyncio.sleep(0.2)
-        self.assertTrue(macro.tasks[1].is_holding())
-        self.assertGreater(len(self.result), 2)
-
-        """up"""
-
-        macro.release_trigger()
-        await asyncio.sleep(0.05)
-        self.assertFalse(macro.tasks[1].is_holding())
-
-        self.assertEqual(self.result[0], (EV_KEY, keyboard_layout.get("1"), 1))
-        self.assertEqual(self.result[-1], (EV_KEY, keyboard_layout.get("3"), 0))
-
-        code_a = keyboard_layout.get("a")
-        self.assertGreater(self.result.count((EV_KEY, code_a, 1)), 2)
-
-        self.assertEqual(self.count_child_macros(macro), 1)
-        self.assertEqual(self.count_tasks(macro), 4)
-
-    async def test_hold_failing_child(self):
-        # if a child macro fails, hold will not try to run it again.
-        # The exception is properly propagated through both `hold`s and the macro
-        # stops. If the code is broken, this test might enter an infinite loop.
-        macro = Parser.parse("hold(hold(key(a)))", self.context, DummyMapping)
-
-        class MyException(Exception):
-            pass
-
-        def f(*_):
-            raise MyException("foo")
-
-        macro.press_trigger()
-        with self.assertRaises(MyException):
-            await macro.run(f)
-
-        await asyncio.sleep(0.1)
-        self.assertFalse(macro.running)
-
-    async def test_dont_hold(self):
-        macro = Parser.parse("key(1).hold(key(a)).key(3)", self.context, DummyMapping)
-
-        asyncio.ensure_future(macro.run(self.handler))
-        await asyncio.sleep(0.2)
-        self.assertFalse(macro.tasks[1].is_holding())
-        # press_trigger was never called, so the macro completes right away
-        # and the child macro of hold is never called.
-        self.assertEqual(len(self.result), 4)
-
-        self.assertEqual(self.result[0], (EV_KEY, keyboard_layout.get("1"), 1))
-        self.assertEqual(self.result[-1], (EV_KEY, keyboard_layout.get("3"), 0))
-
-        self.assertEqual(self.count_child_macros(macro), 1)
-        self.assertEqual(self.count_tasks(macro), 4)
-
-    async def test_just_hold(self):
-        macro = Parser.parse("key(1).hold().key(3)", self.context, DummyMapping)
-
-        """down"""
-
-        macro.press_trigger()
-        asyncio.ensure_future(macro.run(self.handler))
-        await asyncio.sleep(0.1)
-        self.assertTrue(macro.tasks[1].is_holding())
-        self.assertEqual(len(self.result), 2)
-        await asyncio.sleep(0.1)
-        # doesn't do fancy stuff, is blocking until the release
-        self.assertEqual(len(self.result), 2)
-
-        """up"""
-
-        macro.release_trigger()
-        await asyncio.sleep(0.05)
-        self.assertFalse(macro.tasks[1].is_holding())
-        self.assertEqual(len(self.result), 4)
-
-        self.assertEqual(self.result[0], (EV_KEY, keyboard_layout.get("1"), 1))
-        self.assertEqual(self.result[-1], (EV_KEY, keyboard_layout.get("3"), 0))
-
-        self.assertEqual(self.count_child_macros(macro), 0)
-        self.assertEqual(self.count_tasks(macro), 3)
-
-    async def test_dont_just_hold(self):
-        macro = Parser.parse("key(1).hold().key(3)", self.context, DummyMapping)
-
-        asyncio.ensure_future(macro.run(self.handler))
-        await asyncio.sleep(0.1)
-        self.assertFalse(macro.tasks[1].is_holding())
-        # since press_trigger was never called it just does the macro
-        # completely
-        self.assertEqual(len(self.result), 4)
-
-        self.assertEqual(self.result[0], (EV_KEY, keyboard_layout.get("1"), 1))
-        self.assertEqual(self.result[-1], (EV_KEY, keyboard_layout.get("3"), 0))
-
-        self.assertEqual(self.count_child_macros(macro), 0)
-
-    async def test_hold_down(self):
-        # writes down and waits for the up event until the key is released
-        macro = Parser.parse("hold(a)", self.context, DummyMapping)
-        self.assertEqual(self.count_child_macros(macro), 0)
-
-        """down"""
-
-        macro.press_trigger()
-        await asyncio.sleep(0.05)
-        self.assertTrue(macro.tasks[0].is_holding())
-
-        asyncio.ensure_future(macro.run(self.handler))
-        macro.press_trigger()  # redundantly calling doesn't break anything
-        await asyncio.sleep(0.2)
-        self.assertTrue(macro.tasks[0].is_holding())
-        self.assertEqual(len(self.result), 1)
-        self.assertEqual(self.result[0], (EV_KEY, keyboard_layout.get("a"), 1))
-
-        """up"""
-
-        macro.release_trigger()
-        await asyncio.sleep(0.05)
-        self.assertFalse(macro.tasks[0].is_holding())
-
-        self.assertEqual(len(self.result), 2)
-        self.assertEqual(self.result[0], (EV_KEY, keyboard_layout.get("a"), 1))
-        self.assertEqual(self.result[1], (EV_KEY, keyboard_layout.get("a"), 0))
-
-    async def test_aldjfakl(self):
-        repeats = 5
-
-        macro = Parser.parse(
-            f"repeat({repeats}, key(k))",
-            self.context,
-            DummyMapping,
-        )
-
-        self.assertEqual(self.count_child_macros(macro), 1)
 
     async def test_2(self):
         start = time.time()
@@ -694,42 +444,6 @@ class TestMacros(MacroTestBase):
             self.handler
         )
         self.assertEqual(macro_variables.get("e"), "2")
-
-    async def test_multiline_macro_and_comments(self):
-        # the parser is not confused by the code in the comments and can use hashtags
-        # in strings in the actual code
-        comment = '# repeat(1,key(KEY_D)).set(a,"#b")'
-        macro = Parser.parse(
-            f"""
-            {comment}
-            key(KEY_A).{comment}
-            key(KEY_B). {comment}
-            repeat({comment}
-                1, {comment}
-                key(KEY_C){comment}
-            ). {comment}
-            {comment}
-            set(a, "#").{comment}
-            if_eq($a, "#", key(KEY_E), key(KEY_F)) {comment}
-            {comment}
-        """,
-            self.context,
-            DummyMapping,
-        )
-        await macro.run(self.handler)
-        self.assertListEqual(
-            self.result,
-            [
-                (EV_KEY, KEY_A, 1),
-                (EV_KEY, KEY_A, 0),
-                (EV_KEY, KEY_B, 1),
-                (EV_KEY, KEY_B, 0),
-                (EV_KEY, KEY_C, 1),
-                (EV_KEY, KEY_C, 0),
-                (EV_KEY, KEY_E, 1),
-                (EV_KEY, KEY_E, 0),
-            ],
-        )
 
 
 if __name__ == "__main__":
